Fix build on linux x86_64

+++ src/ui/layout/WidgetHierarchyItem.cpp
@@ -70,7 +70,7 @@ void WidgetHierarchyItem::refreshOrderingData(bool resort, bool recursive)
 
     if (!_manipulator || !_manipulator->getWidget()) return;
 
-    setData(_manipulator->getWidgetIndexInParent(), Qt::UserRole + 1);
+    setData(static_cast<uint>(_manipulator->getWidgetIndexInParent()), Qt::UserRole + 1);
 
     if (recursive)
     {
